Fix linting and CI (#119)
First, golangci-lint was failing because linting was executed without building
and embedding assets first, which leaves the app/ package
unbuildable/unlintable. I tried adding just the UI building as a dependency for
the "lint" step, but that wasn't sufficient somehow, but finally adding the
entire "build" target as a dependency worked. Except now yamllint was failing
on YAMLs included in the "node_modules" directory, which I excluded.

In Prometheus we seem to avoid the entire initial problem by having build tags
for the web assets with a fallback Go file to read files from the file system
when not building with "builtinassets". But I didn't want to introduce that
whole shebang here as well... maybe there's a more ideal way to do this though?

Signed-off-by: Julius Volz <julius.volz@gmail.com>